fix: commit21:10:23
This is a long commit body supposed to describe the useful change made.
I know this is not usefull at all but what you know, I have to test this is somehow working.
To be onest I'm very happy up to the point.

BREAKING CHANGE: the cable will not work anymore